don't use `system` to open files
This removes the ability to use `system` to open files. Instead we use `Process.runSync`.

Quoting the [manpage for `system`](https://man.archlinux.org/man/system.3):
> Any user input that is employed as part of command should be carefully sanitized,
> to ensure that unexpected shell commands or command options are not executed.

AFAICT only spaces are replaced with `\ ` right now. This is not enough, and invoking the
`system` function is also completely unnecessary. Creating a new process directly is both
simpler and better. This option was introduced in https://github.com/crazecoder/open_file/pull/127,
however only on linux and not as the default. This is not sufficient.

Running the following in an unsandboxed flutter app on linux would open `gnome-calculator`
(if installed, but you can place any command in there that will be executed):
```dart
  OpenFile.open(";gnome-calculator");
```

I haven't tested this on macOS but it _should_ work there similarly to linux.